% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coeffeasy_logit.R
\name{Coeffeasy_logit}
\alias{Coeffeasy_logit}
\title{Interpreta fácilmente los coeficientes de un modelo logístico.}
\usage{
Coeffeasy_logit(modelo, x = NULL, y = NULL, alfa = 0.05)
}
\arguments{
\item{modelo}{Un objeto de modelo de tipo glm con familia binomial (modelo logístico).}

\item{x}{Nombre de la variable predictora de interés (carácter). Si es NULL, la función intentará identificarla automáticamente.}

\item{y}{Nombre de la variable de respuesta (carácter). Si es NULL, la función intentará identificarla automáticamente.}

\item{alfa}{Nivel de significatividad para pruebas de hipótesis. El valor predeterminado es 0.05.}
}
\value{
Una cadena de texto con la interpretación de los coeficientes del modelo logístico.
}
\description{
Esta función toma un modelo logístico y devuelve una interpretación simplificada
de los coeficientes, especialmente enfocado en el cambio en la probabilidad
asociado con un cambio unitario en la variable predictora.
}
