% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coeffeasy_multinom.R
\name{Coeffeasy_multinom}
\alias{Coeffeasy_multinom}
\title{Easily interpret the coefficients of a multinomial model.}
\usage{
Coeffeasy_multinom(modelo, x = NULL, aplha = 0.05)
}
\arguments{
\item{modelo}{A multinom object (multinomial model).}

\item{x}{Name of the predictor variable of interest (character). If NULL, the function will try to identify it automatically.}

\item{aplha}{Significance level for hypothesis testing. The default value is 0.05.}
}
\value{
A text string with the interpretation of the multinomial model's coefficients.
}
\description{
This function takes a multinomial model and returns a simplified interpretation
of the coefficients, focusing especially on the change in log-odds
associated with a unit change in the predictor variable for each category of the dependent variable.
}
