% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coeffeasy_lm.R
\name{Coeffeasy_lm}
\alias{Coeffeasy_lm}
\title{Interpretation of Coefficients for Linear Regression Models}
\usage{
Coeffeasy_lm(
  model,
  x = NULL,
  y = NULL,
  alpha = 0.05,
  error = "HC2",
  cluster = NULL,
  short = T,
  language = "en"
)
}
\arguments{
\item{model}{An object of class \code{lm} representing the linear regression model.}

\item{x}{A character string specifying the name of the predictor variable.}

\item{y}{A character string specifying the response variable.}

\item{alpha}{A numeric value for the significance level. Default is 0.05.}

\item{error}{A character string specifying the type of heteroskedasticity consistent standard errors. Default is "HC2".}

\item{cluster}{A value to specify the clustering variable. Default is NULL.}

\item{short}{A logical flag indicating if the output message should be shorter and more concise. Default is TRUE.}

\item{language}{A character string indicating the desired output language. Default is "en" (English). Uses Google Translate via the \code{polyglotr} package.}
}
\description{
This function provides an easy interpretation of coefficients from linear regression models.
It also checks for potential heteroskedasticity and uses, by default, the HC2 correction to
the standard errors if it is detected. Additionally, it allows for multilingual output by leveraging the
\code{polyglotr} package to translate the interpretation message.
}
