% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coeffeasy_logit.R
\name{Coeffeasy_logit}
\alias{Coeffeasy_logit}
\title{Easily interpret the coefficients of a logistic model.}
\usage{
Coeffeasy_logit(modelo, x = NULL, y = NULL, alfa = 0.05)
}
\arguments{
\item{modelo}{A glm object with binomial family (logistic model).}

\item{x}{Name of the predictor variable of interest (character). If NULL, the function will try to identify it automatically.}

\item{y}{Name of the response variable (character). If NULL, the function will try to identify it automatically.}

\item{alfa}{Significance level for hypothesis testing. The default value is 0.05.}
}
\value{
A text string with the interpretation of the logistic model's coefficients.
}
\description{
This function takes a logistic model and returns a simplified interpretation
of the coefficients, focusing especially on the change in probability
associated with a unit change in the predictor variable.
}
